+++ lib/src/core/room.dart
@@ -40,6 +40,7 @@ import '../support/disposable.dart';
 import '../support/platform.dart';
 import '../support/region_url_provider.dart';
 import '../support/websocket.dart' show WebSocketException;
+import '../track/audio_management.dart';
 import '../track/local/audio.dart';
 import '../track/local/video.dart';
 import '../track/track.dart';
@@ -263,6 +264,10 @@ class Room extends DisposableChangeNotifier with EventsEmittable<RoomEvent> {
         logger.warning('could not fetch region settings $e');
       }));
     }
+
+    // configure audio for native platform
+    await NativeAudioManagement.start();
+
     try {
       await engine.connect(
         _regionUrl ?? url,
@@ -687,6 +692,14 @@ class Room extends DisposableChangeNotifier with EventsEmittable<RoomEvent> {
     // trigger change notifier only if list of participants membership is changed
     var hasChanged = false;
     for (final info in updates) {
+      // The local participant is not ready yet, waiting for the
+      // `RoomConnectedEvent` to create the local participant.
+      if (_localParticipant == null) {
+        await events.waitFor<RoomConnectedEvent>(
+          duration: const Duration(seconds: 10),
+        );
+      }
+
       if (localParticipant?.identity == info.identity) {
         await localParticipant?.updateFromInfo(info);
         continue;
@@ -943,6 +956,8 @@ extension RoomPrivateMethods on Room {
 
     // clean up engine
     await engine.cleanUp();
+
+    await NativeAudioManagement.stop();
 
     // reset params
     _name = null;