+++ lib/src/participant/local.dart
@@ -215,8 +215,13 @@ class LocalParticipant extends Participant<LocalTrackPublication> {
 
     if (room.engine.enabledPublishCodecs?.isNotEmpty ?? false) {
       // fallback to a supported codec if it is not supported
-      if (!room.engine.enabledPublishCodecs!.any((c) =>
-          publishOptions?.videoCodec == mimeTypeToVideoCodecString(c.mime))) {
+      if (!room.engine.enabledPublishCodecs!
+          .where((c) => c.mime.startsWith('video/'))
+          .where(
+              (c) => videoCodecs.any((v) => c.mime.toLowerCase().endsWith(v)))
+          .any((c) =>
+              publishOptions?.videoCodec ==
+              mimeTypeToVideoCodecString(c.mime))) {
         publishOptions = publishOptions.copyWith(
           videoCodec: mimeTypeToVideoCodecString(
                   room.engine.enabledPublishCodecs![0].mime)
@@ -735,7 +740,6 @@ class LocalParticipant extends Participant<LocalTrackPublication> {
           await publication.mute(stopOnMute: stopOnMute);
         }
       }
-      await room.applyAudioSpeakerSettings();
       return publication;
     } else if (enabled) {
       if (source == TrackSource.camera) {